refactor(SupervisorFilters): implement filters components separetly
