+++ src/locales/en.json
@@ -136,6 +136,9 @@
       },
       "filters": {
         "filter_conversations": "Filter conversations",
+        "filter_conversations_drawer": "Filter conversations",
+        "apply_filters": "Apply filters",
+        "clear_filters": "Clear filters",
         "status": {
           "conversations": "Conversations",
           "in_progress": "In progress",
