+++ src/locales/es.json
@@ -140,6 +140,9 @@
       },
       "filters": {
         "filter_conversations": "Filtrar conversaciones",
+        "filter_conversations_drawer": "Filtrar conversaciones",
+        "apply_filters": "Aplicar filtros",
+        "clear_filters": "Limpiar filtros",
         "status": {
           "conversations": "Conversaciones",
           "in_progress": "En progreso",
