+++ src/locales/pt_br.json
@@ -140,6 +140,9 @@
       },
       "filters": {
         "filter_conversations": "Filtrar conversas",
+        "filter_conversations_drawer": "Filtrar conversas",
+        "apply_filters": "Aplicar filtros",
+        "clear_filters": "Limpar filtros",
         "status": {
           "conversations": "Conversas",
           "in_progress": "Em progresso",
