+++ src/views/AgentBuilder/Supervisor/SupervisorFilters/index.vue
@@ -1,183 +1,60 @@
 <template>
   <section class="supervisor-filters">
-    <!-- <section class="supervisor-filters__selects">
-      <UnnnicInputDatePicker
-        v-model="dateFilter"
-        position="left"
-        class="selects__date-picker"
-        :maxDate="today"
-        data-testid="date-picker"
-      />
-      <UnnnicSelectSmart
-        v-model:modelValue="statusFilter"
-        :options="statusOptions"
-        orderedByIndex
-        multiple
-        multipleWithoutSelectsMessage
-        autocomplete
-      />
-      <UnnnicSelectSmart
-        v-model:modelValue="csatFilter"
-        :options="csatOptions"
-        orderedByIndex
-        multiple
-        multipleWithoutSelectsMessage
-        autocomplete
-      />
-      <UnnnicSelectSmart
-        v-model:modelValue="topicFilter"
-        :options="topicOptions"
-        orderedByIndex
-        multiple
-        multipleWithoutSelectsMessage
-        autocomplete
-      />
-    </section> -->
-
     <FilterText data-testid="filter-text" />
 
     <UnnnicButton
       iconLeft="filter_list"
       type="secondary"
+      :text="$t('agent_builder.supervisor.filters.filter_conversations')"
       data-testid="filter-button"
+      @click="openFilterDrawer"
+    />
+
+    <UnnnicDrawer
+      :modelValue="isFilterDrawerOpen"
+      class="supervisor-filters__drawer"
+      :title="
+        $t('agent_builder.supervisor.filters.filter_conversations_drawer')
+      "
+      :primaryButtonText="$t('agent_builder.supervisor.filters.apply_filters')"
+      :secondaryButtonText="
+        $t('agent_builder.supervisor.filters.clear_filters')
+      "
+      @close="closeFilterDrawerWithReset"
+      @primary-button-click="applyFilters"
+      @secondary-button-click="closeFilterDrawerWithReset"
     >
-      {{ $t('agent_builder.supervisor.filters.filter_conversations') }}
-    </UnnnicButton>
+      <template #content>
+        <FilterDate />
+        <FilterStatus />
+        <FilterCsat />
+        <FilterTopics />
+      </template>
+    </UnnnicDrawer>
   </section>
 </template>
 
 <script setup>
-import { ref, watch, computed } from 'vue';
+import { ref } from 'vue';
 
 import FilterText from './FilterText.vue';
-import { format } from 'date-fns';
-import { useSupervisorStore } from '@/store/Supervisor';
-import i18n from '@/utils/plugins/i18n';
+import FilterDate from './FilterDate.vue';
+import FilterStatus from './FilterStatus.vue';
+import FilterCsat from './FilterCsat.vue';
+import FilterTopics from './FilterTopics.vue';
 
-const supervisorStore = useSupervisorStore();
+const isFilterDrawerOpen = ref(false);
 
-const today = format(new Date(), 'yyyy-MM-dd');
-function getQueryFilterArray(filter, filterOptions) {
-  if (!supervisorStore.filters[filter]) return [];
-  if (typeof supervisorStore.filters[filter] !== 'string')
-    return supervisorStore.filters[filter];
-
-  return supervisorStore.filters[filter].split(',').map((item) => {
-    if (item === '') return { label: '', value: '' };
-
-    return filterOptions.value.find(
-      (option) => option.value === item || option.label === item,
-    );
-  });
+function openFilterDrawer() {
+  isFilterDrawerOpen.value = true;
 }
 
-const dateFilter = ref({
-  start: supervisorStore.filters.start,
-  end: supervisorStore.filters.end,
-});
+function closeFilterDrawerWithReset() {
+  isFilterDrawerOpen.value = false;
+}
 
-watch(
-  () => dateFilter.value,
-  () => {
-    supervisorStore.filters.start = dateFilter.value.start;
-    supervisorStore.filters.end = dateFilter.value.end;
-  },
-  { immediate: true },
-);
-
-const getStatusTranslation = (filter) =>
-  i18n.global.t(`agent_builder.supervisor.filters.status.${filter}`);
-
-const statusOptions = computed(() => [
-  { label: getStatusTranslation('conversations'), value: '' },
-  {
-    label: getStatusTranslation('optimized_resolution'),
-    value: 'optimized_resolution',
-  },
-  {
-    label: getStatusTranslation('other_conclusion'),
-    value: 'other_conclusion',
-  },
-  {
-    label: getStatusTranslation('transferred_to_human_support'),
-    value: 'transferred_to_human_support',
-  },
-  { label: getStatusTranslation('in_progress'), value: 'in_progress' },
-  { label: getStatusTranslation('unclassified'), value: 'unclassified' },
-]);
-const statusFilter = ref(getQueryFilterArray('status', statusOptions));
-
-watch(
-  () => statusFilter.value,
-  () => {
-    supervisorStore.filters.status = statusFilter.value.map(
-      (status) => status?.value || '',
-    );
-  },
-  { immediate: true, deep: true },
-);
-
-const getCsatTranslation = (filter) =>
-  i18n.global.t(`agent_builder.supervisor.filters.csat.${filter}`);
-
-const csatOptions = computed(() => [
-  { label: getCsatTranslation('csat'), value: '' },
-  { label: getCsatTranslation('very_satisfied'), value: 'very_satisfied' },
-  { label: getCsatTranslation('satisfied'), value: 'satisfied' },
-  { label: getCsatTranslation('neutral'), value: 'neutral' },
-  { label: getCsatTranslation('dissatisfied'), value: 'dissatisfied' },
-  {
-    label: getCsatTranslation('very_dissatisfied'),
-    value: 'very_dissatisfied',
-  },
-]);
-const csatFilter = ref(getQueryFilterArray('csat', csatOptions));
-
-watch(
-  () => csatFilter.value,
-  () => {
-    supervisorStore.filters.csat = csatFilter.value.map(
-      (csat) => csat?.value || '',
-    );
-  },
-  { immediate: true, deep: true },
-);
-
-const topicOptions = computed(() => [
-  {
-    label: i18n.global.t(`agent_builder.supervisor.filters.topic.topic`),
-    value: '',
-  },
-]);
-const topicFilter = ref([]);
-
-const isRequestedTopics = ref(false);
-
-watch(
-  () => topicFilter.value,
-  async () => {
-    if (!isRequestedTopics.value) {
-      await getTopics();
-      topicFilter.value = getQueryFilterArray('topics', topicOptions);
-      isRequestedTopics.value = true;
-    }
-    supervisorStore.filters.topics = topicFilter.value.map(
-      (subject) => subject?.label || '',
-    );
-  },
-  { immediate: true, deep: true },
-);
-
-async function getTopics() {
-  await supervisorStore.getTopics().then((topics) => {
-    topicOptions.value = [
-      ...topicOptions.value,
-      ...topics.map((topic) => ({
-        label: topic.name,
-        value: topic.uuid,
-      })),
-    ];
-  });
+function applyFilters() {
+  isFilterDrawerOpen.value = false;
 }
 </script>
 
@@ -187,15 +64,13 @@ async function getTopics() {
   grid-template-columns: 8fr 4fr;
   gap: $unnnic-spacing-sm;
 
-  &__selects {
-    display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    gap: $unnnic-spacing-sm;
+  &__drawer {
+    :deep(.unnnic-drawer__container .unnnic-drawer__content) {
+      overflow: visible;
 
-    .selects__date-picker {
-      :deep(.input) {
-        width: 100%;
-      }
+      display: flex;
+      flex-direction: column;
+      gap: $unnnic-spacing-sm;
     }
   }
 }
